Update CHEF-KOCH's NSABlocklist + excessive GOV URLs.txt
This isn't yet another hosts file or DNSBL that claims to secure the web, it's specially designed to stop known NSA / GCHQ / C.I.A. or F.B.I. servers from being connecting to you without permission, of course the IP's also can be used for Bot Revolt or other tools. The list is not designed to block common malware, spyware/ads or anything that is already available on the net via a proper designed hosts for such special case. This hosts or the super ranges lists could block some of your sites/servers you may need, so you'll be warned!

My list is original based on 2007 published Wikileaks documents and includes my CHEF-KOCH's modifications from 2008, 2012, 2014 and 2015.

This fork:
Removes most of the documents that aren't blocklists
Makes blocklists compatible with pi-hole
Adds a blocklist that removes the excessive GOV URLs (since pi-hole blocks at DNS level)
Updates README (while giving complete credit and donor link to CHEF-KOCH to continue to his work)

This forked project includes:
A 'HOSTS (including excessive GOV URLs)' file that includes all Servers/DNS domains that are known to be involved in spying. The confirmation is given within the Research link(s) at the bottom and with my own tests.